Clean up extra ui models
