+++ solr-admin-app/solr_admin/models/restricted_condition_table.py
@@ -2,8 +2,8 @@
 from . import db
 
 # The class that corresponds to the database table for restricted conditions.
-class RestrictedCondition2(db.Model):
-    __tablename__ = 'restricted_condition2'
+class RestrictedConditionTable(db.Model):
+    __tablename__ = 'restricted_condition'
 
     cnd_id = db.Column(db.Integer, primary_key=True)
     cnd_text = db.Column(db.VARCHAR(1000))
@@ -11,7 +11,6 @@ class RestrictedCondition2(db.Model):
     consent_required = db.Column(db.VARCHAR(1))
     consenting_body = db.Column(db.VARCHAR(195))
     instructions = db.Column(db.VARCHAR(1000))
-    word_phrase = db.Column(db.VARCHAR(300))
